requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudipsp==1.0.3
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.9
django-allauth==0.46.0
django-braces==1.15.0
django-celery-results==2.2.0
django-filter==21.1
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.3.2
flower==1.0.0
humanize==3.12.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
kombu==5.2.2
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
pipenv==2021.11.15
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
pycosat==0.6.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.13.3
